fix the type bug(off_t and size_t is long(4G) < file_size  )and fix relocTail(), pass the test.

+++ skiplist.cc
@@ -108,7 +108,7 @@ std::list<kvpair> skiplist_type::scan(key_type start, key_type end) const {
   }
   return pairs;
 }
-size_t skiplist_type::size() const { return ele_number; }
+uint64_t skiplist_type::size() const { return ele_number; }
 /**
 @brief get all key-value pair list
  * @return std::list<kvpair>
